Add warning if not selected note and you pressing delete or edit button (edit button warning is not working properly so it will be fixed later). Also, add checking for empty input in add and edit windows.

+++ library/app.py
@@ -1,6 +1,6 @@
 from PySide6 import QtWidgets
 from PySide6.QtCore import Qt
-from PySide6.QtWidgets import QMainWindow
+from PySide6.QtWidgets import QMainWindow, QMessageBox
 from PySide6.QtSql import QSqlTableModel
 from MainWindow import Ui_MainWindow
 from EditWindow import Ui_Dialog
@@ -27,10 +27,10 @@ class Notebook(QMainWindow):
         self.model.select()
         self.ui.tableView.setModel(self.model)
         # self.ui.tableView.setColumnHidden(0, True)
-        self.model.setHeaderData(1, Qt.Horizontal, "Название")  # Title
-        self.model.setHeaderData(2, Qt.Horizontal, "Описание")  # Description
-        self.model.setHeaderData(3, Qt.Horizontal, "Дата и время")  # DateTime
-        self.model.setHeaderData(4, Qt.Horizontal, "Уровень")  # Level
+        self.model.setHeaderData(1, Qt.Horizontal, "Название")
+        self.model.setHeaderData(2, Qt.Horizontal, "Описание")
+        self.model.setHeaderData(3, Qt.Horizontal, "Дата и время")
+        self.model.setHeaderData(4, Qt.Horizontal, "Уровень")
         self.ui.tableView.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
 
     def open_add_note_window(self):
@@ -47,7 +47,6 @@ class Notebook(QMainWindow):
 
         self.ui_window.pushButton_3.clicked.connect(self.new_window.close)
         self.ui_window.pushButton_2.clicked.connect(self.reset)
-        # self.ui_window.pushButton.clicked.connect(self.add_note)
 
     def radiobutton(self):
         if self.ui_window.radioButton.isChecked():
@@ -55,8 +54,6 @@ class Notebook(QMainWindow):
         elif self.ui_window.radioButton_2.isChecked():
             self.level = 'Средний'
         elif self.ui_window.radioButton_3.isChecked():
-            self.level = 'Низкий'
-        else:
             self.level = 'Низкий'
 
     def reset(self):
@@ -71,7 +68,11 @@ class Notebook(QMainWindow):
         title = self.ui_window.lineEdit.text()
         description = self.ui_window.lineEdit_2.text()
         datetime = self.ui_window.dateTimeEdit.text()
-        level = self.level # self.ui_window.verticalLayout.currentText()
+        level = self.level
+
+        if not title or not description or not datetime or not level:
+            QMessageBox.warning(self.new_window, "Ошибка", "Все поля должны быть заполнены, а также выбран приоритет.")
+            return
 
         self.conn.add_new_note_query(title, description, datetime, level)
 
@@ -79,15 +80,25 @@ class Notebook(QMainWindow):
         self.new_window.close()
 
     def edit_note(self):
-        self.radiobutton()
+        index = self.ui.tableView.selectedIndexes()
 
-        index = self.ui.tableView.selectedIndexes()[0]
+        if not index:
+            QMessageBox.warning(self.new_window, "Ошибка", "Пожалуйста, выберите запись для изменения.")
+            return  # Доделать
+
+        index = index[0]
         id_ = str(self.ui.tableView.model().data(index))
+
+        self.radiobutton()
 
         title = self.ui_window.lineEdit.text()
         description = self.ui_window.lineEdit_2.text()
         datetime = self.ui_window.dateTimeEdit.text()
-        level = self.level # self.ui_window.verticalLayout.currentText()
+        level = self.level
+
+        if not title or not description or not level:
+            QMessageBox.warning(self.new_window, "Ошибка", "Все поля должны быть заполнены, а также выбран приоритет.")
+            return
 
         self.conn.edit_note_query(title, description, datetime, level, id_)
 
@@ -95,7 +106,14 @@ class Notebook(QMainWindow):
         self.new_window.close()
 
     def delete_note(self):
-        index = self.ui.tableView.selectedIndexes()[0]
+        self.main_window = QtWidgets.QDialog()
+        index = self.ui.tableView.selectedIndexes()
+
+        if not index:
+            QMessageBox.warning(self.main_window, "Ошибка", "Пожалуйста, выберите запись для удаления.")
+            return
+
+        index = index[0]
         id_ = str(self.ui.tableView.model().data(index))
 
         self.conn.delete_note_query(id_)
